Port TestProjectWithAddons, TestPrivateLink
